Add cloth config demo
